Create Longest Path in Tree
INSTRUCTIONS
Write a function that computes the length of the longest path of consecutive integers in a tree. 

A node in the tree has a value and a set of children nodes. A tree has no cycles and each node has exactly one parent. A path where each node has a value 1 greater than its parent is a path of consecutive integers (e.g. 1,2,3 not 1,3,5). 

A few things to clarify:

Integers are all positive
Integers appear only once in the tree