Create histograms_and_areas.cpp
Given an array of non-negative integers that represent the bars (y value) in a histogram (with the array index being the x value), find the rectangle with the largest area under the curve and above the x-axis (i.e. the largest rectangle that fits inside the histogram). Return the pair of array indices that represent the rectangle.

For array [2,4,2,1], the largest area is 6, with height 2, and width from indices 0 to 2
For array [2,4,2,1,10,6,10] the largest area is 18, with height 6 and width from indices 4 to 6